app/requirements.txt
backcall==0.2.0
backports.entry-points-selectable==1.1.0
blis==0.7.5
bunch==1.0.1
catalogue==2.0.6
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cymem==2.0.6
debugpy==1.5.0
decorator==5.1.0
distlib==0.3.3
dnspython==2.1.0
entrypoints==0.3
filelock==3.3.0
Flask==2.0.2
Flask-Cors==3.0.10
fr-core-news-md @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-3.2.0/fr_core_news_md-3.2.0-py3-none-any.whl
geographiclib==1.52
geopy==2.2.0
huggingface-hub==0.4.0
idna==3.3
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jupyter-client==7.0.6
jupyter-core==4.8.1
langcodes==3.3.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
murmurhash==1.0.6
nest-asyncio==1.5.1
nltk==3.6.7
numpy==1.21.2
packaging==21.3
pandas==1.3.3
parso==0.8.2
pathy==0.6.1
pickleshare==0.7.5
Pillow==9.0.0
pipenv==2021.5.29
platformdirs==2.4.0
preshed==3.0.6
prompt-toolkit==3.0.20
pycparser==2.21
pydantic==1.8.2
Pygments==2.10.0
pymongo==3.12.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
regex==2022.1.18
requests==2.27.1
sacremoses==0.0.47
scikit-learn==1.0.1
scipy==1.7.2
sentence-transformers==2.1.0
sentencepiece==0.1.96
six==1.16.0
smart-open==5.2.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
thinc==8.0.13
threadpoolctl==3.0.0
tokenizers==0.11.4
torch==1.10.2
torchvision==0.11.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
transformers==4.16.1
typer==0.4.0
typing_extensions==4.0.1
urllib3==1.26.8
virtualenv==20.8.1
virtualenv-clone==0.5.7
vosk==0.3.32
wasabi==0.9.0
wcwidth==0.2.5
Werkzeug==2.0.2
mpu
langdetect
pyaudioconvert
icecream
